fast

+++ Profile/Network.ps1
@@ -29,7 +29,10 @@ function Switch-VMSwitch-ByConnection {
     )
 
     # Admin kontrolü
-    Assert-AdminRights
+    if ( -not (Assert-AdminRights-Windows) ) {
+        Write-Host "Bu script'in yönetici olarak çalıştırılması gerekiyor!" -ForegroundColor Red
+        return
+    }
 
     # vEthernet adapter IP kontrolü
     $wifiEx = Get-NetIPConfiguration -InterfaceAlias "vEthernet ($WifiSwitch)" -ErrorAction SilentlyContinue
@@ -82,7 +85,11 @@ function Remove-FullTunnel {
     )
 
     try {
-        Assert-AdminRights
+
+        if ( -not (Assert-AdminRights-Windows) ) {
+            Write-Host "Bu script'in yönetici olarak çalıştırılması gerekiyor!" -ForegroundColor Red
+            return
+        }
 
         Write-Host "`n[+] Full-Tunnel rotaları temizleniyor..." -ForegroundColor Yellow
         Remove-NetRoute -DestinationPrefix "128.0.0.0/1" -NextHop $NextHop -Confirm:$false -ErrorAction SilentlyContinue
@@ -102,7 +109,10 @@ function Set-DNS {
     )
 
     try {
-        Assert-AdminRights
+        if ( -not (Assert-AdminRights-Windows) ) {
+            Write-Host "Bu script'in yönetici olarak çalıştırılması gerekiyor!" -ForegroundColor Red
+            return
+        }
         
         Write-Host "[+] DNS sunucusu ayarlanıyor: $NextHop" -ForegroundColor Cyan
         Set-DnsClientServerAddress -InterfaceAlias $InterfaceAlias -ServerAddresses @($NextHop)
